+++ apps/web/app/components/blocks/UtilityBar/forms/UtilityBarActionsForm.vue
@@ -53,7 +53,7 @@
 import { SfSwitch } from '@storefront-ui/vue';
 import draggable from 'vuedraggable/src/vuedraggable';
 import dragIcon from '~/assets/icons/paths/drag.svg';
-import type { ActionType, UtilityBarProps, ActionsSettings, ActionOrderItem } from './types';
+import type { ActionType, UtilityBarProps, ActionsSettings, ActionOrderItem } from '../types';
 
 const { blockUuid } = useSiteConfiguration();
 const route = useRoute();
